msm: dcvs: provide frequency set/get callbacks at registration
There is no need to register a separate structure for setting and
getting frequency.

Simply pass function pointers to set and get callbacks when a
core is registered.

While at it rename the msm_dcvs_freq_sink_register/unregister to
msm_dcvs_freq_sink_start/stop to better reflect that those
apis are meant to do.

Change-Id: I0e5bb1fe8d127e5841ee9916dad5c3fd64815228
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>