drm/exynos: implement an optimized page_flip
We want to avoid all the mode_set logic and just update the
dma_addr when page-flipping. This is necessary for the parallel
page_flip support. We want to avoid a having to call
crtc_update (which can fail) and then call commit. We need
a single call (which we call from the kds callback) which cannot
fail.

BUG=chrome-os-partner:12170
TEST=VT switch. Log in/out. Suspend/resume. Web GL. Youtube. HDMI.

Change-Id: I21afa7dd15bbff184c7ae759bb26e23d855af372
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/32787