drm/exynos: implement an optimized page_flip
We want to avoid all the mode_set logic and just update the
dma_addr when page-flipping. This is necessary for the parallel
page_flip support. We want to avoid a having to call
crtc_update (which can fail) and then call commit. We need
a single call (which we call from the kds callback) which cannot
fail.

BUG=chrome-os-partner:12170
TEST=VT switch. Log in/out. Suspend/resume. Web GL. Youtube. HDMI.

Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34198
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>

(cherry picked from commit 69518a3c650de9186f345a5e7feaddbadfc394d0)

Change-Id: Ia4972c9008b1f8c6a5e1f5f5c3239a575a0fe704
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34312
Reviewed-by: Anush Elangovan <anush@google.com>