CHROMIUM: sbs-battery: probe should try talking to the device
Turns out this driver doesn't actually try talking to the device at probe
time, so if it's incorrectly configured in the device tree or platform
data (or if it's been removed from the system), then probe will succeed
and every later access will sit there and time out. The end result is
a system that thinks it has a battery but can never read status, which
isn't very useful.

Instead, just read any register (I chose status) at boot, and if that
fails, don't register the device.

BUG=chrome-os-partner:11381
TEST=modify device tree to have battery at other address, try booting
and make sure the 2 minute timeout is gone (and dmesg has the error
string in it).

Change-Id: I2b9fcbf5de42ec84fab615edd003cbf37adbc2f9
Original-Change-Id: I728910968a40a55de3b3c9a10921a3684ace215f
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/32406
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/32495