msm: qpnp-power-on: Add API to reset and shutdown the system
Add an API to allow the system to reset or shutdown, depending on
the parameter passed. The qpnp-power-on driver supports any number
of PON devices, but only one device is permissible to reset the
device. Introduce a new Device Tree binding to indicate which
device should be responsible for this resetting the system.

Change-Id: I4086b50402a5c997fb0efada3e65116eec8f4e53
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>