diag: fix hsic concurrency issue when in MEMORY_DEVICE_MODE
When the logging mode changes to MEMORY_DEVICE_MODE, an array
is used to house the pointers to the data read from the hsic
until the data can be copied to user space. Since the storing
and copying actions access this array in different threads,
access to this array can result in concurrent access by these
threads. Note that the storing access to the array happens in
interrupt context and the copying access happens in process
context.  Use spin_locks to protect agains this concurrency
issue.

Also, there is unnecessary queuing of the hsic read work function
on the work queue. Add checks to remove this unnecessary queuing.

(cherry picked from 13c4d01a12c0206a68124623930c58a7c0f3ff06)

Conflicts:

	drivers/char/diag/diagchar_core.c
	drivers/char/diag/diagfwd_hsic.c

Change-Id: I215ae699e6b56a38abfc9fb51b7cb082b31bfb00
CRs-fixed: 373328
Signed-off-by: Dixon Peterson <dixonp@codeaurora.org>
Signed-off-by: Jessica Gonzalez <jgaona@codeaurora.org>