msm: smd: call smem_find() after releasing spinlocks
If a subsystem crashes with a remote spinlock locked, then the subsystem
restart code should release it before attempting to access SMEM.
Originally when this code was written, smem_find() did not lock a remote
spinlock in violation of the SMEM protocol.  Once smem_find() was
updated to lock the remote spinlock, this deadlock scnario was created.

Do not call smem_find() until the remote spinlocks have been released
for the failed subsystem.

CRs-Fixed: 397287
Change-Id: I41086c81df5943659465741601191be9d243ca2c
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>