UPSTREAM: mwifiex: use country ie of requested AP while associating
AP's country ie is parsed and 11d information is updated in firmware
while associating. Previous mwifiex_send_domain_info_cmd_fw() call in
association path is not necessary now. Hence it is removed in this
patch.

Signed-off-by: Amitkumar Karwar <akarwar@marvell.com>
Signed-off-by: Bing Zhao <bzhao@marvell.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>

BUG=chrome-os-partner:12711
TEST=Enabled dynamic debug to dump the cmd 0x5b being sent to f/w
and verified the cmd body contains the country ie for the AP we are
associating.

Change-Id: Ia3d764433daf22bd8d3b90a3cfc49d2904189870
Reviewed-on: https://gerrit.chromium.org/gerrit/33099
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Reviewed-by: Gary Morain <gmorain@chromium.org>
Commit-Ready: Bing Zhao <bzhao@marvell.com>
Tested-by: Bing Zhao <bzhao@marvell.com>