FIXUP: ARM: dma-mapping: remove custom consistent dma region
This attempts to fix the above comment to match more closely
what landed upstream.  We originally had v3 of the patch but
it went up to v6.

There are still a few differences from upstream.  Most notably:
* I didn't take the VM_ARM_DMA_CONSISTENT stuff.  I wasn't sure
  how far reaching that was.
* We have a slightly different way for disabling CMA.  I didn't
  try to match upstream.

Note that part of the original patch moved to a separate
patch: "add more sanity checks in arm_dma_mmap()".  See the next
CL.

BUG=chrome-os-partner:13935
TEST=Bootup

Change-Id: I9a642884fbd3dcb17f68e687539a7772caf41f8c
Signed-off-by: Marek Szyprowski <m.szyprowski@samsung.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/32961
Reviewed-by: Pawel Osciak <posciak@google.com>