thermal: tsens8974: Add support to read redundant calibration
Add support to read the redundant region of the QFPROM for
TSENS calibration. An efused bit is used to determine if
the redundant QFPROM calibration region is to be read
instead of the primary calibration region. This supports
cases when there is an error while programming data to
the primary TSENS calibration region.

The following change updates the devicetree bindings for
the QFPROM calibration region size to map the entire
region to include the back-up calibration region as well.

Change-Id: I773786516a0149d5bb16f8c72c777921f6e14aa4
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>