CHROMIUM: iommu/exynos: Fix allocation to be safer in terms of alignment
The code was expecting kzalloc to align to the size of the allocation
even though that isn't guaranteed, and even had a BUG() to catch when it
happened.  This started triggering with SLUB debugging turned on.  Fix it
by using a slab which returns objects that are properly aligned.

BUG=chrome-os-partner:13948
TEST=boot with slub debugging turn on

Change-Id: Icdf9475233e8c80f090f624a4ca5497011d4aa33
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/33053
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Pawel Osciak <posciak@google.com>
Reviewed-by: Prathyush Kalashwaram <prathyush@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>