msm: mpq8064: VC report field info that is used by VP
VP assumed that the first field provided was a bottom field, if VP
was incorrect in this assumption it would not properly de-interlace
content. It was also possible for VC to drop a field, when this
happened VP would become out of sync again causing de-interlacing
to be incorrect.

CRs-Fixed: 397534
Change-Id: I7a75a5dcd27149220273284a3d48f407bc5a4f31
Signed-off-by: Terence Hampson <thampson@codeaurora.org>