CHROMIUM: config: Modify generic x86_64 config
This change adds intel idle support, and sets the number of cpus
to the kernel default value of 64.

Intel idle knows which processors it will work correctly on, and on
systems it is not aware of, it will fall back to acpi idle.  Enabling
it here should improve robustness and power on systems intel idle
supports, and should not change anything on systems it doesn't support.

The number of CPU's setting was previously 4.  Setting this value too
low can have serious power consequences (non-initialized cores will
prevent the entire system from entering sleep states) and the additional
overhead is minimal. A value of 64 should be large enough to cover the
majority of systems available that Chromium OS might be used with.

BUG=None
TEST=Smoke tests passed on amd64-generic and amd64-corei7 using x86_64
     flavour.

Change-Id: I3385d751913fa1f4c81cc0b0d9fb9a69519d7f15
Signed-off-by: Russell Webb <russell.webb@intel.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/33315
Reviewed-by: paul drews <paul.drews@intel.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>