mmc: msm sdcc: Enabling the "Cache Ctrl" feautre
Cache is a temporary storage space in eMMC device.
The cache should in typical case reduce the access time (compared to an
access to the main non-volatile storage)
for both write and read. The cache is not directly accessible by the
host. Data in the cache may (and most probably will) be lost due to a
sudden power down. If there was a flush operation ongoing when the
power was lost then also any such data may be lost.
Turning the cache OFF shall trigger flushing of the data to the
non-volatile storage.

(cherry picked from commit 19e21065991090f12adda9716885bf6b283cd41c)

Change-Id: I79f6bca78267c1916b6808184570392a5431af1a
Signed-off-by: Yaniv Gardi <ygardi@codeaurora.org>