Revert "EXYNOS: USB: Modify Common block power-down control for USB 2.0 PHY"
This reverts commit e62200853145dfa26d7614a0004b4a4912a901ae

The above commit allegedly causes problems with shutdown / suspend
on some systems, especially when running a recovery image and when
nothing is plugged into the USB 2.0 port.  It also has been known to
cause problems with no_console_suspend.  Although we don't
understand the connection between that CL and the problems,
it's wiser to revert while we investigate.

Nothing in this CL is so critical that a revert will really hurt.

This does the same thing as
<https://gerrit.chromium.org/gerrit/#/c/30814/>, but with a commit
message that indicates that it's a straight revert.

BUG=chrome-os-partner:10932, chrome-os-partner:13068
TEST=Automatic reboot after recovery works with this CL.

Change-Id: Ic01b4494b9e4414d22424c010280757d6e15e5a2
Reviewed-on: https://gerrit.chromium.org/gerrit/33400
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Katie Roberts-Hoffman <katierh@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>