msm: camera: disable unwanted interrupts from vfe
Why: For stopping the VFE we are halting AXI
and we poll on status register to see if halting
is completed. Once the halting is completed we
were not clearing the interrupt from interrupt
status register causing halt to be processed again.
With this change we are clearing halt interrupt
bits from status register.

CRs-Fixed: 386273

Change-Id: Ia0712de4691bf64ec8529fef88f766b1e72f16b1
Signed-off-by: Roja Rani Yarubandi <rojay@codeaurora.org>