mmc: msm_sdcc: Reset interrupt MASK register (MCI_MASK0) on powering off
In one corner case, the SD card is stuck in a bad state with its DAT0
line pulled low, and SDCC is waiting on the interrupt when the line
goes back high. But due its bad state eventually the SD card is removed
from the system.

Later during re-scanning of the devices, the SD card is power-cycled and
added to the system. But now the pending interrupts for SD card is
received as the interrupt MASK register (MCI_MASK0) was not cleared.

To prevent such cases reset the interrupt MASK register (MCI_MASK0) while
powering off to prevent any pending interrupts after power-cycle.

CRs-Fixed: 396706
Change-Id: I83c087a76cd92f20c5990a787296e093a13478ad
Signed-off-by: Pratibhasagar V <pratibha@codeaurora.org>