android/lowmemorykiller: Only consider gfp_mask free pages
In certain memory configurations there can be a large number of
CMA pages which are not suitable to satisfy certain memory
requests.
This large number of unsuitable pages can cause the
lowmemorykiller to not kill any tasks because the
lowmemorykiller counts all free pages.

In order to ensure the lowmemorykiller properly evaluates the
free memory only count the free pages which are suitable for
satisfying the memory request.

(cherry picked from commit cc4baf70665f54305182bfaf98c6236d129e6dad)

Change-Id: Ic11f3d658bd143b1e0bf1a76f419c0a2ca33caea
Signed-off-by: Liam Mark <lmark@codeaurora.org>
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>