tty: smux_ctl: Do not set POLLERR if signal received
During the userspace freeze process, a fake signal is sent to userspace
applications.  In the case of SMUX CTL, the smux_ctl_poll() will set the
POLLERR flag if a signal is received.  This incorrectly unblocks the
userspace application that called poll() and the userspace application
will grab a wakelock before doing additional processing which aborts the
freeze process.

Change smux_ctl_poll() to not set POLLERR if a signal is received.

(cherry picked from commit 3cbcfa3d79564063fc4aa52173feeff101926cb9)

Change-Id: Ifbc479c18b2add4b5afa2cfac312e6358143fe3d
CRs-Fixed: 402651
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>