tty: smux_ctl: Do not set POLLERR if signal received
During the userspace freeze process, a fake signal is sent to userspace
applications.  In the case of SMUX CTL, the smux_ctl_poll() will set the
POLLERR flag if a signal is received.  This incorrectly unblocks the
userspace application that called poll() and the userspace application
will grab a wakelock before doing additional processing which aborts the
freeze process.

Change smux_ctl_poll() to not set POLLERR if a signal is received.

(cherry picked from commit 3cbcfa3d79564063fc4aa52173feeff101926cb9)

Change-Id: I42e11447eceed4889d0246abd0c49ba926cf1505
CRs-Fixed: 402651
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>
Signed-off-by: Jessica Gonzalez <jgaona@codeaurora.org>