msm: SSR: Add an API to mark subsystems as "already online"
Some subsystems have not moved to use the subsystem_get()/put()
API so we don't have proper tracking of their state.  Add an API
to mark these subsystems as "already online", so that if they
crash we know that they really crashed for a reason. This API
should be removed once the affected subsystems move to the
get/put APIs.

Change-Id: I597dfc8c1f7c6259a1c6fc6204d28a81835472b5
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>