CHROMIUM: Revert "CHROMIUM: Input: cyapa - Disable runtime PM"
This reverts commit f1256eb8c33081f160a66b5d4ac63e3c094ff558.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:14583,chrome-os-partner:10550, chrome-os-partner:12954
TEST=Boot a system with cyapa trackpad.
cd /sys/bus/i2c/devices/1-0067/power/
(the i2c bus will vary per system above. 1-0067 is for snow. 7-0067 is
typical for Parrot or Lumpy)
check that runtime_suspend_scanrate_ms exists. Cat it.
Check that it returns 30.
cat runtime_status
Check that it is "suspended" when the touchpad has not been touched
for 2 seconds. Check that it is "active" when the touchpad is being touched
or for 2 seconds after last touch.

Change-Id: Iaabe26da2f188d238329ebc3ba5695c9663bf95f
Reviewed-on: https://gerrit.chromium.org/gerrit/34063
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>