staging:iio:adis16400: Fix and cleanup 3db filter setting
First of all the 3db divisors table is partially wrong and incomplete, secondly
the code would round up to the next higher frequency if the requested frequency
would exactly match one of the available frequencies and finally it would reject
frequencies larger than the largest supportable frequency. This patch fixes the
first two issues and modifies the code to round down to the largest supportable
frequency if a frequency higher than that is requested.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>