msm: mdm: Add shutdown ioctl and send poweroff request to the mdm
When the whole phone is being powered off, a poweroff request needs
to be sent to the external modem so that it can shutdown gracefully.
This request needs to be triggered from userspace before kernel drivers
start unloading so that other drivers needed to send the request are
still available. The shutdown ioctl is provided for this purpose.
The request is sent over system monitor.

Since the mdm driver is common to platforms that support sysmon as
well as those that don't, a platform specific flag named send_shdn
is added for runtime detection.

Crs-Fixed: 401598
Signed-off-by: Joel King <joelking@codeaurora.org>

Conflicts:

	arch/arm/mach-msm/include/mach/mdm2.h
	arch/arm/mach-msm/mdm_common.c
	arch/arm/mach-msm/mdm_private.h
	include/linux/msm_charm.h

Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>
(cherry picked from commit c060882e9d724c595b27f8937062302760b13ee6)

Change-Id: Ieb47cd327f49a484e699ff612d6c9e691003a5e8
Signed-off-by: Sivasri Kumar Vanka <sivasri@codeaurora.org>