msm: mdm: Add shutdown ioctl and send poweroff request to the mdm
When the whole phone is being powered off, a poweroff request needs
to be sent to the external modem so that it can shutdown gracefully.
This request needs to be triggered from userspace before kernel drivers
start unloading so that other drivers needed to send the request are
still available. The shutdown ioctl is provided for this purpose.
The request is sent over system monitor.

Since the mdm driver is common to platforms that support sysmon as
well as those that don't, a platform specific flag named send_shdn
is added for runtime detection.

(cherry picked from commit bf3e4b526c55ea765a92f54cf91540330d0b75ae)

Change-Id: I356200b6e407b3fbe7a9db20675cdd2fbcf1bd78
Crs-Fixed: 401598
Signed-off-by: Joel King <joelking@codeaurora.org>