msm: mdm: Add shutdown ioctl and send poweroff request to the mdm
When the whole phone is being powered off, a poweroff request needs
to be sent to the external modem so that it can shutdown gracefully.
This request needs to be triggered from userspace before kernel drivers
start unloading so that other drivers needed to send the request are
still available. The shutdown ioctl is provided for this purpose.
The request is sent over system monitor.

Since the mdm driver is common to platforms that support sysmon as
well as those that don't, a platform specific flag named send_shdn
is added for runtime detection.

Crs-Fixed: 401598
Signed-off-by: Joel King <joelking@codeaurora.org>

Conflicts:

	arch/arm/mach-msm/include/mach/mdm2.h
	arch/arm/mach-msm/mdm_common.c
	arch/arm/mach-msm/mdm_private.h
	include/linux/msm_charm.h

Change-Id: I36de5ce98e14a92a7b816208190a373f6f1b3f65
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>