thermal: tsens8960: Add platform detection
TSENS is not supported on the MSM8960 1.0 because of a
HW bug. The driver used a platform detection mechanism
through a flag. With the 8960PRO the CPU MAJOR
version resets back to major version 1. The driver
does not distinguish the PRO parts differently and
disables the driver. Change the method to detect
the CPU using the platform detection api instead
of relying on the platform data.

(cherry picked from commit 30447fec965309e30f1eb33c0d318ee4a5365026)

Change-Id: Ib3a0b13f0e1cf6e8a41b2ac48666df6cbd709a81
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>