msm: acpuclock-8974: Populate missing user_vco_mask in hfpll_data
This mask is needed in order to switch VCO modes when running the
PLL at frequencies above 1700MHz. Specify the missing data. Because
rates about 1.7GHz are not used yet, the missing mask should not
have been the source of any real problems.

Also remove the TODO line that was inserted as a reminder to
double-check the 'magic' PLL configuration register values after
hardware characterization. That has since been done and the
original values were determined to be fine.

Change-Id: I548c02e44cc5fded740e513252110e7318fec6e8
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>