KDS: Don't declare wait queue on stack then pass pointers to it
The kds_waitall() function was declaring a wait queue on the stack and
then passing pointers around to it.  This is not a good idea unless
you're incredibly careful that nobody could possibly ever use the wait
queue after your function exits.

In the case of kds_waitall(), there could have been two problems:

1. When kds_waitall() returned it could return rset which has a
   pointer to this on-stack memory in rset->callback_parameter.
   It's unclear that anyone would use this, but it seems like a bad
   idea.
2. Since no locking / synchronization is done in the "!triggered"
   case, it's possible that we'll observe "rset->pending == 0" before
   the sync_cb() gets called.  That could mean the function exiting
   before the sync_cb() gets called and the sync_cb() reading/trashing
   random stack memory.  This is similar to a problem observed in the
   adc driver as described in:
     https://gerrit.chromium.org/gerrit/#/c/34223/

I have not done a full analysis to see if the above is happing or
causing any problems but it seems wise to move the wait queue off the
stack to preempt any possible problems.

Placed the wait queue in the kds_resource_set so that we can be sure
that it's scope is exactly the same as the pointer to it--it will get
freed at the exact same time.

BUG=chrome-os-partner:14665
TEST=Watch a youtube video with no problems.

Change-Id: Iaff3028f67a0c9a540e4b9bd989e6649840433e8
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34455
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Pawel Osciak <posciak@google.com>