power: qpnp-charger: add VCP workaround
The voltage collapse protection feature is broken
in hardware on PM8941 2.0.

Fix this by adding a workaround which overrides
the ICHG_END_VCP_CMP to 1 before setting the
maximum input current limit for the USB charging
path.

Note that this workaround does not apply to the
DC charging path.

Also add flags to track which workarounds should be
enabled or disabling specific based on the runtime-detected
PMIC version.

Change-Id: Ie18e7cd3f70e4896fa5c04e952cfbc5c5b3dc6ef
Signed-off-by: David Keitel <dkeitel@codeaurora.org>