arm/dts: Add support for PMIC PM8644 regulator devices
The new dtsi files are arranged so that parameters can be
overridden in convenient ways. msm-pm8644.dtsi contains the platform
independent characteristics for each of the PMIC regulator devices.
This includes the SPMI base register for each regulator device,
its name, its compatible string, and its status.

The next file, mpq8092-regulator.dtsi, then fills in platform
dependent regulator configurations for the 8092 platform.  This
includes min/max voltage constraints, parent regulator, and
various qpnp-regulator parameters.

The value of status is deliberately overridden between
msm-pm8644.dtsi and mpq8092-regulator.dtsi so
that devices are only added for regulators which should be
controlled via the qpnp-regulator driver. When parsing a file
from top to bottom, the last value specified for a given parameter
is the one that ends up in the dtb file. Only the actively used
devices will be configured via mpq8092-regulator.dtsi
with status="okay".

Change-Id: Ib958cfaba25695197fc57d670888817489a58b26
Signed-off-by: Ravi Kumar V <kumarrav@codeaurora.org>