mmc: msm_sdcc: don't return error if regulator info is not available
On some platforms, few SDCC slots may be powered by always-on regulators
on board and in this case, regulator info for such SDCC slot is not
defined as part of their platform data. So let's not return error from
driver if regulator info is not present for a SDCC slot.

Signed-off-by: Asutosh Das <asutoshd@codeaurora.org>
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>
(cherry picked from commit d5902bf929db15820a1242561a984a67bcc27944)

Change-Id: Ibe5e7af22e29ff416202186dcc3f0cab51ba52d1
Signed-off-by: Gopal Goberu <ggober@codeaurora.org>