CHROMIUM: ALSA: ASoC: exynos hdmi - always set switch state.
Don't drop the enable/disable switch toggle if it is called when HDMI is
not connected.  If it is dropped, user space has to call it with timing
constraints that it can't know.  Let the switch toggle if hotplug isn't
set, audio will be configured when hotplug is detected.

BUG=chrome-os-partner:14435
TEST=toggle HDMI switch and check that it is set.

Change-Id: I6c553cf4ada19616783ceffe111459b046d01b6f
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34579
Reviewed-by: Rahul Sharma <rahul.sharma@samsung.com>
Tested-by: Rahul Sharma <rahul.sharma@samsung.com>
(cherry picked from commit 9fc0875583130bd69346d918580f6a1c9a30394d)
Reviewed-on: https://gerrit.chromium.org/gerrit/34924