qseecom:Fix race condition while voting for clocks
In the current functionality, if DFAB clock is already on and a
request is made to turn on SFPB clock, the DFAB clock is turned
off and SFPB clock is turned on and vice-versa.

The above situation can lead to unexpected errors. The current fix
makes sure that clocks that are on are not voted to be off without
explicit request.

(cherry picked from commit 8d7565828aca6023ca968d2e5350737bd6d86215)

Change-Id: I52f525b1b459f2f7f028cd01867c4747b4bbc0ec
Signed-off-by: Ramesh Masavarapu <rameshm@codeaurora.org>