defconfig: msm8974: Disable CONFIG_DEBUG_MUTEXES
Disable mutex debugging on 8974, due to an apparent deadlock
issue on the spinlocks this debugging mechanism uses. This
can be re-enabled once the root cause of the issue is identified
and addressed.  Until then, it should remain disabled due to
the high number of watchdog barks resulting from the deadlock.

Change-Id: If5a2a4a186d4c08625b4fabb3875e0370c1a0678
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>