staging:iio:adis16400: Remove samplerate_available attribute
While the samplerate supported by the devices supported by this driver is not
continuous it supports a wide range, much more than currently listed in the
samplerate_available attribute. Also it accept all values written to the
samplerate attribute and will round-up them to the nearest supported sample
rate. So remove the samplerate_available attribute.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>