CHROMIUM: mwifiex: Increase channel dwell time
Increase channel dwell time for active-scan channels (which also
happen to be the channels that tend to have the most congestion).
In congested channels, we have tested this to improve the chance
of scan success by over a factor of 3.  Scan time is still less
than 7 seconds which is still faster than some mac80211-based
chipsets.

Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chrome-os-partner:14655
TEST=Scan loops, interop test

Change-Id: I2e279d6f6e9719bbf2d7317f139ef58431fe92a2
Reviewed-on: https://gerrit.chromium.org/gerrit/34838
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>