drm/exynos: flip_apply the correct buffer
From the initial report:

<snip>
There's an issue with the queueing up swaps with future_kds_extra.
When we do so, we still immediately write the fb that we should flip
to into the crtc (in exynos_drm_crtc_page_flip). The KDS callback
displays whatever is the fb currently stored there
(exynos_drm_kds_callback -> exynos_drm_crtc_page_flip_apply).

So when we get a second flip request before the first has completed,
we overwrite the fb that the first flip was intended to flip to.
This breaks the relationship between the KDS resource that we wait
for and the buffer that we scan out. This clearly has the potential
to cause tearing.
</snip>

Fix is to flip_apply the fb that we are getting the callback for.

BUG=chrome-os-partner:14195,chrome-os-partner:14940
TEST=Verified that WebGL aquarium is fixed.

Change-Id: I87a1e3ff60437da92a68266f81fc7bcd6288542c
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/34974