CHROMIUM: block: prevent race
In block/genhd.c in disk_clear_events, there was a race that could cause the
clearing flag in the disk_events structure to be changed before we use it. The
race was when another call to disk_events_workfn (made from a workqueue), runs
after we call __disk_unblock_events and before we take the spin lock.

To fix this, I changed the function __disk_unblock_events to not take the
spinlock in disk_events. There is now a requirement to hold the spinlock
whenever you make a call to that function.

This allows us to eliminate the window in which another call to
disk_events_workfn could run since we are already holding the spinlock when we
unblock events.

TEST=suspend and resume with usb device plugged in many times
BUG=chrome-os-partner:14892, chromium-os:35228

Change-Id: I3658553d150dadee27b5575d1b73fe76999aa5d2
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/35378
Reviewed-by: Paul Taysom <taysom@chromium.org>