msm: kgsl: Log retired timestamps on device wake
Knowing the retired timestamp on wake-up is useful to debug hangs that
are caused due to clock instability. When we request for clock ON, the
clock may/may-not be turned ON, but we go ahead assuming it is ON. This
can be caught by printing the timestamp on wake. The timestamp we print,
can be compared against the timestamp in postmortem dump (assuming a
hang happens).

If there is a difference, we can say that GPU has run for sometime after
wake. So, we needn't suspect clocks.

If there is no difference, we can say that GPU has hung immediately after
wake. So, we can positively suspect the clocks.

(cherry picked from commit d0348f1644a21ae73982f94c442bdf9f17584ebe)

(cherry picked from commit 0184185f274762462738621282dc50815313339d)

(cherry picked from commit 9e5af757c0902544113bba8ac5296d4a5f15aaf7)

Change-Id: Ie871fecf6fcf71098648ac583e3890669b1e4a68
Signed-off-by: Ranjhith Kalisamy <ranjhith@codeaurora.org>
Signed-off-by: Gopal Goberu <ggober@codeaurora.org>
Signed-off-by: Gopal Goberu <ggober@codeaurora.org>