asoc: hdmi: removed unwanted error messages from hdmi audio plugin
This patch removes unwanted error messages from hdmi-audio plugin.
These messages are printed when hdmi is not plugged in and plugin
functions are called for hdmi audio configuration. This is done for
caching the audio parameters in the hdmi plugin for later use and
not an error situation.

BUG=None
TEST=verified, no print comes.

Change-Id: Idb06520a6e982a6d329d9b963596f1236e090e67
Signed-off-by: Rahul Sharma <rahul.sharma@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/35402
Tested-by: Shirish S <shirish@chromium.org>
Reviewed-by: Prashanth Godrehal <prashanth.g@samsung.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Ready: Shirish S <shirish@chromium.org>