arm/dt: msm8974-liquid: Add SMB210 5 V regulator device
Add a device node for MSM8974 LIQUID to provide control of the
SMB210 regulator which supplies 5 V to various onboard devices.

The SMB210 regulator (named "ext_5v") is controlled via PM8941
MPP 2.  Add initial configurations for this MPP so that it can
be used to successfully enable and disable the regulator.

Modify the parent-supply property of regulators pm8941_mvs1 and
pm8941_mvs2 to point at the ext_5v regulator.  This is needed
because these regulators are supplied by ext_5v on MSM8974 LIQUID
as opposed to by pm8941_boost on other MSM8974 boards.

Change-Id: I62bee4eaafb4508521b547f1e1aea4f2fb4fae53
Signed-off-by: David Collins <collinsd@codeaurora.org>