power: qpnp-charger: add API to enable and disable charging
The power supply framework exposes an API which allows to
disable and enable charging.

Add this feature to the qpnp-charger driver to allow
other drivers to enable and disable charging.

When charging is disabled and a battery is present the charger
is forced to run of the battery even if a charger adapter
is connected.

Change-Id: I7be4a4022de5614a7f803a0377182fc6529c1bd8
Signed-off-by: David Keitel <dkeitel@codeaurora.org>