CHROMIUM: Input: cyapa - Disable irq before changing power modes
There could be a potential race condition here as the irq may
be active as we are trying to read and write to device registers
to change power modes. In order to resolve this, disable irq
before cyapa_set_power_mode in cyapa_suspend.

BUG=chromium-os:35663
TEST=Builds clean.
Check that suspend with power/wakeup enabled and disabled
both work correctly to disable and enable the touchpad wake.
1. echo disabled > /sys/bus/i2c/devices/1-0067/power/wakeup
2. powerd_suspend
3. check that trackpad cannot wake the system.
4. Wake using keyboard. Check mouse is functional.
5. echo enabled > /sys/bus/i2c/devices/1-0067/power/wakeup
6. powerd_suspend
7. check that trackpad can wake the system.
8. Check that the trackpad is operational after resume

Change-Id: Ifd8715c51ddea138c5c4c28b4a4337af8d6021cc
Signed-off-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/36409