ARM: exynos5440_defconfig: add defconfig for EXYNOS5440
Since current exynos_defconfig cannot support for EXYNOS5440 because
of LPAE, there were discussions about that in mailing list. So this
adds initial defconfig which is tested on SSDK5440 and SD5v1 for
LPAE enabled EXYNOS5440. This defconfig enables ARCH_EXYNOS, ARM_LPAE,
PCI, CPU_FREQ, THERMAL, SATA, GMAC and so on for EXYNOS5440.

Signed-off-by: Kukjin Kim <kgene.kim@samsung.com>