msm: mdm: add board level support for an external modem with its own flash
The apq8064 has previously had support for an external mdm9x15 external
modem that does not have its own storage for program images. This new
modem has on-board flash storage.

The gpios used to control power up and reset of this modem are pmic gpios
as opposed to tlmm gpios as previously used.

Change-Id: Id31924e56ccbd579713423b280999dd4909f4aed
Signed-off-by: Joel King <joelking@codeaurora.org>