CHROMIUM: gobi: Keep struct qcusbnet alive while cdev is alive
This fixes an issue where struct qcusbnet is freed while the embedded
character device is still alive.

This uses the ability of cdev to pin an arbitrary kobject by setting
cdev.kobj.parent. We have to switch from kref to kobject to use that.
This is arguably misuse of kobject, but char_dev requires it.t t

BUG=chrome-os-partner:15849
TEST=suspend_stress_test

Change-Id: I309e80c6ec91e491d30ea0bfcb062ae7e55f7198
Signed-off-by: Michael Spang <spang@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/37135
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>