USB: msm_otg: Acknowledge PMIC BSV events, if ID is set
Currently not all the PMIC BSV events are handled due to check
that was introduced for handling spurious interrupts in Host Mode.
Rather than checking the OTG state as apart of these spurious interrupts
which gets updated in workqueue context, check for ID state and return
only, if ID is clear. This will fix all the corner cases and race
conditions as part of handling ACA use cases as well.

CRs-Fixed: 411036
Change-Id: I942b4cf2a71dcdc1fcde785da04fd2b3e1560a6c
Signed-off-by: Chiranjeevi Velempati <cvelempa@codeaurora.org>