CHROMIUM: platform: rename nvram to vbc
As we are actually accessing vboot context and vboot context is not
limited to be stored on nvram, it is clearer to just call it vbc.

This change also removes obsolete comments and makes API interface
consistent (all read/write should return number of bytes instead of zero
on success).

BUG=none
TEST=manual, add debugging sysfs entry to read/write vboot context and
     make sure it works.

Change-Id: Ifd097a1a8af9edb776eed83fe47cb332f8a3445d
Signed-off-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/37686
Reviewed-by: Simon Glass <sjg@chromium.org>