ASoC: mdm9615: Fix the issue of first aplay not working
If the I2S GPIOs are requested before they are installed,
the aplay will not work. The GPIOs were installed in the
prepare function, so aplay worked the second time. The
fix is to move the GPIO install to the dai link init
function, which ensures they will be installed before
they are requested.

CRs-fixed: 425014
Change-Id: Ifb72f65d94e7f21a2d9fc98c4c347672896dc571
Signed-off-by: Damir Didjusto <damird@codeaurora.org>