ASoC: msm: Fix for a seekbar issue during audio playback.
Issue is that when seek happens several times during audio playback,
and at some point last buffer write done event from DSP is received,
an ioctl drain is issued, which will set pending_buffer flag to 0,
In the mean time before drain finishes a seek comes, which issues
a flush on all the buffers and try to resume the playback again.
Now because the pending_buffer flag is set to 0, CMD_RUN will return
without any further writes to DSP. As a result no audio buffers are
reaching the DSP and hence no audio is getting played.

Make a fix is to set pending_buffer flag to 1 whenever playback is resumed
after seek.

CRs-fixed: 416274
Change-Id: I297a3cf0ed41baa1b581da8a314d7929bc5e26f9
Signed-off-by: Harmandeep Singh <hsingh@codeaurora.org>