ASoC: msm: Register restart api for underrun issue
Update LPA platform driver to register restart callback.
The implementation of the callback is to issue a pending buffer
written by the application to restart the session when the dequeue
process is stopped due to underrun issue.

The callback is triggered from ALSA core when the framework
detects that the renderer stopped pulling data but the application
is still writing data. It runs in atomic context.

Signed-off-by: Satya Krishna Pindiproli <satyak@codeaurora.org>
(cherry picked from commit 4650ec00a97c86dc232a27d1b44652c427022e4e)

Change-Id: I62dd56d12b72376caf0526cb33c8940a00c0be4f
Signed-off-by: Sivasri Kumar Vanka <sivasri@codeaurora.org>