ASoC: msm: Register restart api for underrun issue
Update LPA platform driver to register restart callback.
The implementation of the callback is to issue a pending buffer
written by the application to restart the session when the dequeue
process is stopped due to underrun issue.

The callback is triggered from ALSA core when the framework
detects that the renderer stopped pulling data but the application
is still writing data. It runs in atomic context.

Signed-off-by: Satya Krishna Pindiproli <satyak@codeaurora.org>
(cherry picked from commit 0a85bb88426e1c6165dd0504bd3629ba6d5b2a31)

Change-Id: I673c0a556b441793b6cf4616aaaa9075029b1b58
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>