CHROMIUM: vmx: control vmx correctly on resume
On x86 chromeos systems with vmx support, the feature is controlled by
the disablevmx flag. There are several boot paths for the CPUs
and this patch ensures that the resume path for the BSP is covered.

Rename cpu_disable_vmx to cpu_control_vmx. It's a bit more
descriptive since it does not disable vmx in all cases.
Further, have the function itself test the variable; this
is more in keeping with how other such capabilities
are controlled anyway and it keeps the variable in
file scope. Finally, make it all conditional on
CONFIG_CHROMEOS.

We use the variable to control the enable/disable instead
of saved CPU context. I could add the FEATURES MSR
to the saved context and use that on restore but
that adds more changes, and I've seen no evidence that anyone
outside chromeos is interested in disabling VMX, so it
makes more sense to me to limit the footprint of this
patch.

BUG=chrome-os-partner:10449
TEST=apply and boot with this patch.
dmesg | grep Disabling.VMX
Note that all four CPUs have it disabled.
Also
rdmsr 0 0x3a
and see the value.
Suspend any way you want. (powerd_suspend, close the lid, whatever).
Repeat and notice the Disabling message appears again,
for all four cores this time,
and the 0x3a MSRs all have the correct value.

Change-Id: Ie0160db31a1a058ce4d3e4f8752cd0e5d696ea84
[sonnyrao: 3.8: trivial fixups for context change]
Signed-off-by: Ronald G. Minnich <rminnich@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38235
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>

Conflicts:

	arch/x86/kernel/cpu/common.c

[benzh: 3.14 rebase. Resolved trivial conflicts]
Signed-off-by: Ben Zhang <benzh@chromium.org>