qseecom: Check if app exists before loading the app
There is a chance of send a request to QSEE to load an app that
is already previously loaded. Requesting QSEE to load an app
that is already loaded will result in a failure.

Issuing a request to QSEE to check if app already exists before
issuing a request to load it, all within the same app_access_lock
mutex, fixes this issue.

CRs-Fixed: 420473
Signed-off-by: Mona Hossain <mhossain@codeaurora.org>
Signed-off-by: Sivasri Kumar Vanka <sivasri@codeaurora.org>
(cherry picked from commit 27036fecf61ed13332b84d5585ba052fb12fca09)

Change-Id: I1b86bfc6c88c0713dbd6531977c3f46164c0e6b5
Signed-off-by: Sivasri Kumar Vanka <sivasri@codeaurora.org>