power: pm8921-charger: speed up AICL
An issue is observed where it takes 3 to 4 seconds for the AICL
algorithm to settle on a certain USB current. This should be as
fast as possible.

Fix this by circumventing the delay of the unplug check worker if the
USB input current limit has increased.

Also remove the rounding of jiffies to the nearest second
when unplug check worker is called. This way we guarantee
that unplug check worker (and AICL) runs within the timing
requirements.

Conflicts:

	drivers/power/pm8921-charger.c

Change-Id: I546ff57d76e50ad86eb5853782fdcc1f348c9f0e
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
(cherry picked from commit b73001a94665e69f895cbc967104a544abefffa6)