regulator: tps65090: Retry FET enable in case of failure
Sometimes the FET fails to enable but instead gives a timeout error. This
may be a fault with the TPS65090. Turning off the FET and trying again seems
to clear the fault, so do this.

BUG=chrome-os-partner:16222
TEST=manual
Build and boot kernel using a firmware which does not enable the backlight.
See that LCD backlight turns on.

Also need to test on a unit which enhibits the 'no backlight on resume'
problem.

Change-Id: Idb5d867d32b4a290c6afd3d326e8c8f8678588d3
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38679
Reviewed-by: Michael Spang <spang@chromium.org>
Tested-by: Michael Spang <spang@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
(cherry picked from commit d09326be3545c18fb4052e773aaf957be3fd492a)
Reviewed-on: https://gerrit.chromium.org/gerrit/39532
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>