CHERRY-PICK: CHROMIUM: ALSA: hda/ca0132: Set effect settings during initialization.
We did not restore the effect settings after resume, so the hardware
used default values for them. This caused "Surround" to be turned on
after resume.

Now we always set the effect settings in ca0132_init(), so the settings
in hardware and in kernel are in sync.

This patch is provided by Creative.

BUG=chrome-os-partner:16197
TEST=check the output from headphone.

Signed-off-by: Chih-Chung Chang <chihchung@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38719
Reviewed-by: Dylan Reid <dgreid@chromium.org>
(cherry picked from commit 913ba997fe2e78726cdde6824f8a388e81f6ff0a)

Change-Id: I8eca99c588cec43b28ba6b1b0dfa6b916c295ae8
Reviewed-on: https://gerrit.chromium.org/gerrit/41331
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>