CHERRY-PICK: samsung: snow: bitfix: Bail if any pages in a bad chunk were skipped
If we didn't store recovery info for any of the pages in a bad chunk
then we'll consider it an unrecoverable error and reboot.  The old
code only checked the first page in the chunk.  With 8K chunks this
was only a 50% hit rate.

We could just choose to recover the page whose recovery info we have.
We don't do that because there is often more than one bit error in a
chunk and it seems better to error on the side of rebooting.

BUG=chrome-os-partner:15655
TEST=suspend_stress_test

Change-Id: I638cc2f6f5295f4f1dbb5dc68c4b60413da2fbb0
Original-Change-Id:  Ic578bcfaf19efa33f0cc118c9a2f5ca64e5063bb
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Jon Kliegman <kliegs@chromium.org>
(cherry picked from commit f2ad12fd16e1a4378db35f1e567b6d31bf3dbffe)

Reviewed-on: https://gerrit.chromium.org/gerrit/38815