CHROMIUM: drm/exynos: fix refcounting in hdmi (mixer) path
Previously, if two fbs were ready to be flipped during the same
vblank, the kds_callback would drop the frame which came earlier
and call apply on the later frame. This worked fine on fimd but
caused problems on hdmi (mixer). The mixer will only accept one
apply call per vblank and ignore calls after that.

Since, we already applied the earlier frame, the mixer would
ignore the later frame. So the mixer flips the earlier fb while
the refcounting code thinks we are flipping the later fb. So the
ref for the earlier fb gets dropped even though its the scanout
fb. This can cause an IOMMU page fault.

The fix is to hold the later fb in an extra slot and then flip
to it on the next vblank.

In order, to allow calling apply from finish_page_flip, I had to
re-structure mixer_irq_handler in order to avoid spinlock
recursion on reg_slock.

BUG=chrome-os-partner:15241
TEST=see below

Before:

$ DISPLAY=:0.0 xrandr --output eDP-1 --off
* crash *

After:

export DISPLAY=:0.0
for i in `seq 20`; do
  xrandr --output eDP-1 --off;
  sleep 5;
  xrandr --output eDP-1 --auto;
  sleep 5;
done

* no crash *

Change-Id: I45cdad3dea94c9018e6a5fef9f9807b225130cea
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39386
Reviewed-by: Sean Paul <seanpaul@chromium.org>