media: video: Allow NR threshold to be updated by motion est
VCAP did not allow for client to update threshold values based
on the motion estimation. By enabling an option for this threshold
to be updated dynamically by hardware, this allows for a more
aggressive noise reduction to be applied at the request of the
client.

CRs-Fixed: 430970
Signed-off-by: Terence Hampson <thampson@codeaurora.org>
(cherry picked from commit 6a744e339817d68e2b371fa35b0040382b38d044)

Change-Id: Ibc3b6e219382f674a878175b2391368e9298dc7a
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>