msm: ssr: Add subsystem_crashed() API
Add a new API, subsystem_crashed(), so that subsystem users can
indicate that the subsystem has crashed without requiring a
restart. If this API is called, don't restart the subsystem,
instead take note that the call was made and perform a ramdump of
the subsystem once the final subsystem_put() is called.

This allows the video driver to ramdump their subsystem without
requiring them to reboot their subsystem to collect ramdumps.

Change-Id: Iebc514213da3fee47a8ca1be53e6d25602260175
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>