CHROMIUM: ALSA: hda/ca0132 - Limit mic gain.
Hack to set a minimum capture gain of 0 dB (index 90).  This should
never be carried forward as we need to find the root cause of getting
stuck with lower volumes.  For now this is the only link-specific
place to make the change and it fixes the often occurring issue with
the microphone not working in hangouts/apprtc.

BUG=chrome-os-partner:16865
TEST=do several hangouts make sure mixer volume for capture never
drops below 0dB, check that output volume can still move freely.

Change-Id: I1cd9b383dec61aec87cdc4d59fc001005f0700dd
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40033