power: pm8921-charger: set USB OVP time to maximum for PM8917
The USB OVP debounce time play a significant role when a
watchdog reset occurs on PM8917. If the default value of 4
ms is set this can lead the SMBC to mistakenly enter restart_pon
and reset all regulators.

Fix this by increasing the USB OVP debounce time to the maximum
setting of 15ms.

CRs-Fixed: 434656
Change-Id: Iaab82ea254444bba233cd9ebad93d38426154ca4
Signed-off-by: David Keitel <dkeitel@codeaurora.org>