rpc_pmapp: Do not set endpoint variable to NULL
The endpoint varaible is being set to NULL on recieving an error
response in a RPC Call. Due to this RPC client tries to re-register
with the server and causing a memory leak in SMEM. Hence do not set
the Endpoint variable to null on an error response.

(cherry picked from commit 0668d45e1a5ebdfac987011cbe8ffe303e37b651)

Change-Id: I72d1c20fd7a5544de40344a877042a50ef88dcbd
CRs-Fixed: 435094
Signed-off-by: Prasad Sodagudi <psodagud@codeaurora.org>