media: dvb: mpq: Unify PES and section pipes
For each TSIF two pipes were allocated, one for section output
and one for PES output. The two pipes were unified into a single
pipe as the pipes were configured with the same parameters and there's
no real need for separating them. A unified pipe allows recording of
section packets along with non-section packets while maintaining the
order of the packets in the output to be the same as they
appear in the input.

Signed-off-by: Hamad Kadmany <hkadmany@codeaurora.org>
(cherry picked from commit 881d7a23246a2333ed04cd046f837434ced6d817)

Change-Id: If6ea0fc8f21a4ddb2eadac5a2332c6750619c5cf
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>