media: dvb: mpq: Unify PES and section pipes
For each TSIF two pipes were allocated, one for section output
and one for PES output. The two pipes were unified into a single
pipe as the pipes were configured with the same parameters and there's
no real need for separating them. A unified pipe allows recording of
section packets along with non-section packets while maintaining the
order of the packets in the output to be the same as they
appear in the input.

Change-Id: I7ad02cb0f3f8f72b68b4e84e21b1f25801028ff5
Signed-off-by: Hamad Kadmany <hkadmany@codeaurora.org>