CHROMIUM: drm/exynos: fb: Don't take reference when mapping fb's gem objects
The fb itself already takes a reference for all of its gem objects when
it is created, and releases them last when it is destroyed.  Taking an
additional reference when the gem objects are mapped into device memory,
which only happens when the fb is created, seems redundant.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium-os:37097
TEST=builds clean; sanity check on device

Change-Id: I63643e974a8cf7d04d32b96bebd5dad4cc281ba8
Reviewed-on: https://gerrit.chromium.org/gerrit/40448
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>