mmc: core: Don't hold wake lock while detect work is in progress
It is not necessary to hold wake lock and block suspend while
MMC detect work is in progress. If a card is inserted during
device suspend, wait for card to be detected in pm notifiers
before cancelling the detect work. If a card state change is
detected, hold wake lock for 500ms during which userspace be
able to respond to media change events.

(cherry picked from commit 1a37be16272493c51646343ddd54cc4defd69b1f)

Change-Id: I2417b04c26ca6055329abf128b8aa60c782e967f
Signed-off-by: Sujit Reddy Thumma <sthumma@codeaurora.org>