msm7627a: bluetooth: handle BT On failure case proprely
If Bluetooth on is failed due to some reason, clean up is not
proper.
Also when BT turn on fail, we are trying to release some resources
twice. Once as a part of handling the fail cleanup and other
when the userspace initiates turn off process because of turn
on timeout.

CRs-fixed: 439129
Signed-off-by: Ram Mohan Korukonda <rkorukon@codeaurora.org>
(cherry picked from commit b8c882ffcfac4d5cd0fb6e1a9742cb197a835448)

Conflicts:

	arch/arm/mach-msm/board-msm7627a-bt.c

Change-Id: Ib4383553aea5ca8923ab1ccd0a9ed335d1c22243
Signed-off-by: Akhila Musunuri <makhila@codeaurora.org>