gpu: ion: use the VM_MIXEDMAP flag for cached allocations
As of 4b6e1e3702 ("mm: kill vma flag VM_INSERTPAGE" (not yet present
in our tree)), it is required to set VM_MIXEDMAP in order to use
vm_insert_page from a page-fault handler. This flag indicates that a
mapping can contain memory with or without "struct page" backing.

Set the flag to future-proof for 3.7 onwards.

Change-Id: Ia97a84f0cb2d5bf2d6cfef6c27eb71ebe7c60d32
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>