Revert "CHROMIUM: Log RTM_NEWLINK messages broadcast from the kernel (temporary)"
This reverts commit 0a06f97c7e44bbac8944732d7739ea95b79b2ac5

We have gathered all we need to know from this. The problem was in shill
and we have changed that to properly crash instead of silently fail. The
kernel netlink stack seems to work fine as it is and we don't need this
output anymore.

BUG=chromium-os:35479
TEST=Boot the kernel, make sure it still works fine and does not output
any messages with the string "RTM_NEWLINK" to dmesg anymore.

Change-Id: I9a12e41bb25adaf4ef7a7159d97adebfe2fd14e6
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41139
Reviewed-by: Sameer Nanda <snanda@chromium.org>