CHROMIUM: media: tuners: init fw and hw ids to 0 in xc4000.c
In xc4000.c's check_firmware() function, these variables are later
initialized by passing them to xc_get_version() as pointers.  However,
this results in a compile error that these are not initialized since it
is not explicit.

This patch initializes them to zero.  These may or may not be valid
values, but these variables are not used unless xc_get_version() returns
successfully so their initial values aren't important anyway.

I sent this upstream.  It may or may not get accepted, but the directory
structure has changed, so this patch won't apply cleanly when we move to
kernel-next.

BUG=chromium-os:5542
TEST=emerge chromeos-kernel, make sure xc4000.c has no warnings.

Change-Id: I66592ecfb09e36a692329e3ee5c3d1cb3394a2ae
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41473
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>