ASoC: msm: Fix for memory map regions when buffers are not contiguous
This change fixes the issue of sending memory map command
with single region when the buffers allocated for playback
or capture are not contiguous. When the buffers are not
contiguous, we should populate the mregions structure with
all the distinct buffers allocated. Sending a single region
with the total buffer size (bufsz*bufcnt) might overlap with
already mapped region.

Change-Id: I8b1b71748285ab4c1ef278890390cb1e0170a47c
CRs-Fixed: 423082
Signed-off-by: Phani Kumar Uppalapati <phanik@codeaurora.org>