msm: bam_dmux: revert the skb state when memory allocation fails
bam_dmux_write() function modifies the state of a skb via skb_push(), but
if a later allocation fails, the state of the skb is not reverted to the
pre skb_push() state.

The error returned to the caller as a result of the allocation failure gets
propagated to the TCP/IP framework, which re-schedules a retry of the skb
at a later time.  if enough retries occur due to allocation failures,
causes the assertion in skb_push() to fail.

If a later allocation fails, revert the skb buffer state to pre skb_push()
by calling skb_pull().

(cherry picked from commit 381cd549fd0176c649d1579f56b7504aa94df9be)

Change-Id: I11a7b32c33e4967b54886a53565b1da0aae1c347
CRs-Fixed: 441277
Signed-off-by: Arun Kumar Neelakantam <aneela@codeaurora.org>