msm: pm-8x60: Support to do PC at the running core frequency
New version of SAW2 (>v2.1), support turning off/on HFPLL only if the
PLL was active when entering power collapse. This eliminates the need
to switch to a lower frequency running off global PLL, and then turn
off the HFPLL. Power collapse with RPM notification should be a tad
bit faster.

Change-Id: I3014acf1c796a70a3f3d09f84532500e42d727f8
Signed-off-by: Praveen Chidambaram <pchidamb@codeaurora.org>