msm: pil-q6v5: Don't disable Q6 core clock and rclk gating
Clock gating was previously disabled by default so that it wouldn't
interfere with use of JTAG debuggers for the Q6 core. The intention
was to have the Q6 software enable gating for power saving only if
a debugger was not attached.

Instead of requiring the Q6 software to handle this additional
step, leave clock gating enabled by default and rely on debugger
scripts to disable it prior to attaching JTAG to the core.

This provides power saving by default with less software complexity
than the previous solution (which was never fully implemented on the
Q6).

CRs-Fixed: 413190
Change-Id: If9b36d8bb4f8c070e35cd30fa421b8568c13874f
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>